♻️ refactor: switch to postgresql (#1)

+++ migrations/20230327191846_whispers.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS "whispers" (
 	"name" TEXT,
 	"message" TEXT NOT NULL,
 	"private" BOOLEAN NOT NULL,
-	"snowflake" INTEGER NOT NULL UNIQUE,
+	"snowflake" BIGINT NOT NULL UNIQUE,
 	"timestamp" TEXT NOT NULL,
 	PRIMARY KEY ("snowflake")
 );
